+++ JetMETCorrections/Modules/plugins/QGLikelihoodSystematicsESProducer.cc
@@ -9,14 +9,36 @@
 // Original Author:  Salvatore Rappoccio
 //         Created:  Thu, 13 Mar 2014 15:02:39 GMT
 
-#include "JetMETCorrections/Modules/interface/QGLikelihoodSystematicsESProducer.h"
+// system include files
+#include <memory>
+
+// user include files
+#include "FWCore/Framework/interface/ModuleFactory.h"
+#include "FWCore/Framework/interface/EventSetupRecordIntervalFinder.h"
+#include "FWCore/Framework/interface/ESProducer.h"
+#include "FWCore/Framework/interface/ModuleFactory.h"
+
+#include "CondFormats/JetMETObjects/interface/QGLikelihoodObject.h"
+#include "CondFormats/DataRecord/interface/QGLikelihoodSystematicsRcd.h"
+
+class QGLikelihoodSystematicsESProducer : public edm::ESProducer {
+public:
+  QGLikelihoodSystematicsESProducer(const edm::ParameterSet &);
+  ~QGLikelihoodSystematicsESProducer() override{};
+
+  std::unique_ptr<QGLikelihoodSystematicsObject> produce(const QGLikelihoodSystematicsRcd &);
+  void setIntervalFor(const edm::eventsetup::EventSetupRecordKey &, const edm::IOVSyncValue &, edm::ValidityInterval &);
+
+private:
+  edm::ESGetToken<QGLikelihoodSystematicsObject, QGLikelihoodSystematicsRcd> token_;
+};
 
 QGLikelihoodSystematicsESProducer::QGLikelihoodSystematicsESProducer(const edm::ParameterSet& iConfig) {
   //the following line is needed to tell the framework what
   // data is being produced
   std::string label(iConfig.getParameter<std::string>("@module_label"));
-  mAlgo = iConfig.getParameter<std::string>("algo");
-  setWhatProduced(this, label);
+  auto algo = iConfig.getParameter<std::string>("algo");
+  setWhatProduced(this, label).setConsumes(token_, edm::ESInputTag{"", algo});
 }
 
 // The same PDF's is valid for any time
@@ -29,8 +51,7 @@ void QGLikelihoodSystematicsESProducer::setIntervalFor(const edm::eventsetup::Ev
 // Produce the data
 std::unique_ptr<QGLikelihoodSystematicsObject> QGLikelihoodSystematicsESProducer::produce(
     const QGLikelihoodSystematicsRcd& iRecord) {
-  edm::ESHandle<QGLikelihoodSystematicsObject> qglObj;
-  iRecord.get(mAlgo, qglObj);
-
-  return std::make_unique<QGLikelihoodSystematicsObject>(*qglObj);
+  return std::make_unique<QGLikelihoodSystematicsObject>(iRecord.get(token_));
 }
+
+DEFINE_FWK_EVENTSETUP_MODULE(QGLikelihoodSystematicsESProducer);
